Prediction of Production & Query Answering using NLP
Prediction of Production required area as the input and temperature also for the moment. Later on temperature of 20 days after the current date would be taken by web scraping and hence predict the production. The dataset called as querydataset.csv contains questions and answers and then when sentsim.py is executed and a query is given it would return the right answer